+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Unfreeze the first five selection menus for reminder fees. Refs UICIRC-1002, UICIRC-1005, UICIRC-1006.
 * UI tests replacement with RTL/Jest for `RequestCancellationReasons.js`. Refs UICIRC-959.
 * Change functionality for accordions on Circulation settings page (create/edit). Refs UICIRC-1010.
+* Add "Print only" option to patron notice templates. Refs ICIRC-985.
 * Cover Widget by jest tests. Refs UICIRC-899.
 
 ## [9.0.3](https://github.com/folio-org/ui-circulation/tree/v9.0.3) (2023-11-09)